fix sorting and run
